compositor-headless: do not create a seat
The headless backend was creating a seat, but nothing would ever happen
on that seat, except the pointer cursor would show up.

Tests do not seem to use this seat either. weston-test.so plugin creates
its own seat to be controlled by the tests.

This is part of the series to get rid of cursors for screenshot-based
tests where they are not wanted.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.co.uk>
Reviewed-by: Derek Foreman <derekf@osg.samsung.com>